[fix] fix load32

+++ src/bus/clint.rs
@@ -102,10 +102,10 @@ impl Device for Clint {
 
     fn load32(&self, addr: u64) -> Result<u64, (Option<u64>, TrapCause, String)> {
         let addr = self.addr2index(addr);
-        Ok((self.clint[addr + 3] as u64) << 24
-            | (self.clint[addr + 2] as u64) << 16
-            | (self.clint[addr + 1] as u64) << 8
-            | (self.clint[addr + 0] as u64))
+        Ok(((self.clint[addr + 3] as i32) << 24
+            | (self.clint[addr + 2] as i32) << 16
+            | (self.clint[addr + 1] as i32) << 8
+            | (self.clint[addr + 0] as i32)) as i64 as u64)
     }
 
     fn load64(&self, addr: u64) -> Result<u64, (Option<u64>, TrapCause, String)> {
